remove node primitive use polkadex primitive
