--- node-primitives/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "node-primitives"
-version = "2.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-license = "Apache-2.0"
-homepage = "https://substrate.dev"
-repository = "https://github.com/paritytech/substrate/"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
-frame-system = { default-features = false, version = '3.0.0', git = "https://github.com/paritytech/substrate" }
-sp-core = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate" }
-sp-runtime = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate" }
-sp-application-crypto = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate" }
-
-[dev-dependencies]
-sp-serializer = { version = "3.0.0", git = "https://github.com/paritytech/substrate" }
-pretty_assertions = "0.6.1"
-
-[features]
-default = ["std"]
-std = [
-    "codec/std",
-    "frame-system/std",
-    "sp-application-crypto/std",
-    "sp-core/std",
-    "sp-runtime/std",
-]
