+++ node/Cargo.toml
@@ -82,7 +82,7 @@ frame-system = { version = "3.0.0", default-features = false, git = "https://git
 pallet-transaction-payment = { default-features = false, version = '3.0.0', git = "https://github.com/paritytech/substrate" }
 
 #Primitives
-node-primitives = { path = "../node-primitives" }
+polkadex-primitives = { path = "../primitives" }
 sp-authority-discovery = { version = '3.0.0', git = "https://github.com/paritytech/substrate" }
 
 [dev-dependencies]
