+++ node/src/chain_spec.rs
@@ -7,7 +7,7 @@ use node_polkadex_runtime::{
     SessionConfig, SessionKeys, StakerStatus, StakingConfig, SudoConfig, SystemConfig,
     TechnicalCommitteeConfig, MAX_NOMINATIONS,
 };
-use node_primitives::Block;
+use polkadex_primitives::Block;
 use pallet_im_online::sr25519::AuthorityId as ImOnlineId;
 use sc_chain_spec::ChainSpecExtension;
 use sc_service::ChainType;
@@ -22,7 +22,7 @@ use sp_runtime::{
 };
 
 pub use node_polkadex_runtime::GenesisConfig;
-pub use node_primitives::{AccountId, Balance, Signature};
+pub use polkadex_primitives::{AccountId, Balance, Signature};
 
 type AccountPublic = <Signature as Verify>::Signer;
