+++ node/src/service.rs
@@ -613,7 +613,7 @@ mod tests {
     use codec::Encode;
     use node_polkadex_runtime::constants::{currency::CENTS, time::SLOT_DURATION};
     use node_polkadex_runtime::{Address, BalancesCall, Call, UncheckedExtrinsic};
-    use node_primitives::{Block, DigestItem, Signature};
+    use polkadex_primitives::{Block, DigestItem, Signature};
     use sc_client_api::BlockBackend;
     use sc_consensus_babe::{BabeIntermediate, CompatibleDigestItem, INTERMEDIATE_KEY};
     use sc_consensus_epochs::descendent_query;
