+++ primitives/Cargo.toml
@@ -7,16 +7,24 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false }
+codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
 serde = { version = "1.0.101", optional = true }
-sp-core = { git = "https://github.com/paritytech/substrate", default-features = false }
-sp-runtime = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate.git" }
+frame-system = { default-features = false, version = '3.0.0', git = "https://github.com/paritytech/substrate" }
+sp-core = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate" }
+sp-runtime = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate" }
+sp-application-crypto = { version = "3.0.0", default-features = false, git = "https://github.com/paritytech/substrate" }
+
+[dev-dependencies]
+sp-serializer = { version = "3.0.0", git = "https://github.com/paritytech/substrate" }
+pretty_assertions = "0.6.1"
 
 [features]
 default = ["std"]
 std = [
-    "serde",
     "codec/std",
+    "serde",
+    "frame-system/std",
+    "sp-application-crypto/std",
     "sp-core/std",
-    "sp-runtime/std"
+    "sp-runtime/std",
 ]
