+++ runtime/src/constants.rs
@@ -19,7 +19,7 @@
 
 /// Money matters.
 pub mod currency {
-    use node_primitives::Balance;
+    use polkadex_primitives::Balance;
 
     pub const MILLICENTS: Balance = 1_000_000_000;
     pub const CENTS: Balance = 1_000 * MILLICENTS;
@@ -33,7 +33,7 @@ pub mod currency {
 
 /// Time.
 pub mod time {
-    use node_primitives::{BlockNumber, Moment};
+    use polkadex_primitives::{BlockNumber, Moment};
 
     /// Since BABE is probabilistic this is the average expected block time that
     /// we are targeting. Blocks will be produced at a minimum duration defined
